+++ lwk_wollet/src/tx_builder.rs
@@ -540,9 +540,10 @@ impl TxBuilder {
     }
 
     /// Sets the address to drain excess L-BTC to
-    pub fn drain_lbtc_to(mut self, address: Address) -> Self {
+    pub fn drain_lbtc_to(mut self, address: Address) -> Result<Self, Error> {
+        validate_address(&address.to_string(), self.network())?;
         self.drain_to = Some(address);
-        self
+        Ok(self)
     }
 
     /// Adds external UTXOs
@@ -1882,11 +1883,11 @@ impl<'a> WolletTxBuilder<'a> {
     }
 
     /// Wrapper of [`TxBuilder::drain_lbtc_to()`]
-    pub fn drain_lbtc_to(self, address: Address) -> Self {
-        Self {
+    pub fn drain_lbtc_to(self, address: Address) -> Result<Self, Error> {
+        Ok(Self {
             wollet: self.wollet,
-            inner: self.inner.drain_lbtc_to(address),
-        }
+            inner: self.inner.drain_lbtc_to(address)?,
+        })
     }
 
     /// Wrapper of [`TxBuilder::add_external_utxos()`]